requirements.txt
Flask==3.1.0
flask-cors==5.0.1
Flask-JWT-Extended==4.7.1
Flask-Mail==0.10.0
Flask-MySQLdb==2.0.0
mysql-connector-python==9.1.0
mysqlclient==2.2.7
PyJWT==2.10.1
python-dotenv==1.0.1
reportlab==4.2.5
pillow==11.1.0